Move query validation from package org.molgenis.data.validation to org.molgenis.data.validation.query
