+++ molgenis-data-platform/src/main/java/org/molgenis/data/platform/decorators/MolgenisRepositoryDecoratorFactory.java
@@ -17,11 +17,11 @@ import org.molgenis.data.listeners.EntityListenersService;
 import org.molgenis.data.settings.AppSettings;
 import org.molgenis.data.transaction.TransactionInformation;
 import org.molgenis.data.transaction.TransactionalRepositoryDecorator;
-import org.molgenis.data.validation.QueryValidationRepositoryDecorator;
-import org.molgenis.data.validation.QueryValidator;
 import org.molgenis.data.validation.RepositoryValidationDecorator;
 import org.molgenis.data.validation.data.DefaultValueReferenceValidator;
 import org.molgenis.data.validation.data.EntityAttributesValidator;
+import org.molgenis.data.validation.query.QueryValidationRepositoryDecorator;
+import org.molgenis.data.validation.query.QueryValidator;
 import org.molgenis.security.owned.OwnedEntityRepositoryDecorator;
 import org.molgenis.util.EntityUtils;
 import org.springframework.stereotype.Component;
