+++ molgenis-data-validation/src/test/java/org/molgenis/data/validation/query/QueryValidatorTest.java
@@ -1,10 +1,11 @@
-package org.molgenis.data.validation;
+package org.molgenis.data.validation.query;
 
 import org.molgenis.data.*;
 import org.molgenis.data.meta.AttributeType;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.QueryImpl;
+import org.molgenis.data.validation.MolgenisValidationException;
 import org.molgenis.file.model.FileMeta;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.DataProvider;
